requirements.txt
tensorflow-gpu>=2.3.0,<3.0.0
scikit-learn~=0.24.2
tqdm~=4.62.2
wandb~=0.12.3
numpy~=1.20.3